refactore: simplify chart generation as simple function
