+++ apps/admin/package.json
@@ -20,11 +20,8 @@
         "@t3-oss/env-nextjs": "^0.8.0",
         "@tanstack/react-table": "^8.11.7",
         "bcryptjs": "^2.4.3",
-        "canvas": "^2.11.2",
-        "chart.js": "^4.4.1",
         "clsx": "^2.1.0",
         "date-fns": "^3.3.1",
-        "jsdom": "^24.0.0",
         "lucide-react": "^0.320.0",
         "next": "14.1.0",
         "next-auth": "5.0.0-beta.5",
@@ -41,7 +38,6 @@
         "@energyleaf/eslint-config": "workspace:*",
         "@energyleaf/tsconfig": "workspace:*",
         "@types/bcryptjs": "^2.4.6",
-        "@types/jsdom": "^21.1.6",
         "@types/node": "^20.11.15",
         "@types/react": "^18.2.51",
         "@types/react-dom": "^18.2.18",
