+++ packages/ui/package.json
@@ -16,6 +16,7 @@
         "@energyleaf/eslint-config": "workspace:*",
         "@energyleaf/tsconfig": "workspace:*",
         "@turbo/gen": "^1.12.2",
+        "@types/jsdom": "^21.1.6",
         "@types/node": "^20.11.15",
         "@types/react": "^18.2.51",
         "@types/react-dom": "^18.2.18",
@@ -52,9 +53,12 @@
         "@radix-ui/react-toggle": "^1.0.3",
         "@radix-ui/react-tooltip": "^1.0.7",
         "@tanstack/react-table": "^8.11.7",
+        "canvas": "^2.11.2",
+        "chart.js": "^4.4.1",
         "class-variance-authority": "^0.7.0",
         "clsx": "^2.1.0",
         "date-fns": "^3.3.1",
+        "jsdom": "^24.0.0",
         "lucide-react": "^0.320.0",
         "next-themes": "^0.2.1",
         "react-day-picker": "^8.10.0",
